quality_security_platform/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-filter==22.1
django-environ==0.9.0
psycopg2-binary==2.9.5
mysqlclient==2.1.1
django-redis==5.2.0
redis==4.5.1
celery==5.2.7
flower==1.2.0
drf-yasg==1.21.5
networkx==3.1
pyvis==0.3.1
gunicorn==20.1.0
requests==2.28.2
django-crontab==0.7.1
python-json-logger==2.0.4
